refactor: split stack-reconciler.ts into focused modules
Break the 2,150-line StackReconciler orchestrator into eight cohesive
modules so each responsibility can be reasoned about and tested in
isolation. StackReconciler keeps its public surface (plan/apply/update/
stopStack/destroyStack) and becomes a thin wiring layer over the new
managers and handlers.

- stack-diff-generator.ts: pure field-diff + reason helpers
- stack-conflict-detector.ts: port/name conflict detection
- stack-applied-snapshot.ts: Prisma -> lastAppliedSnapshot serialisation
- stack-deployment-logger.ts: failure recording
- stack-infra-resource-manager.ts: Docker network + InfraResource lifecycle
- stack-state-machine-context.ts: HAProxy context build + adopted-web cleanup
- stack-plan-computer.ts: plan/diff computation
- stack-service-handlers.ts: stateful/stateless-web/adopted-web handlers

Net effect: stack-reconciler.ts drops from 2,150 to 747 lines.
